@@ -1,47 +1,42 @@
-import * as React from 'react';
+import {Component, createElement} from 'react';
 import {Animated} from 'react-native';
 
-function createAnimationDongle(className: string) {
-  return React.createClass<any, any>({
-    displayName: 'Cycle' + className,
+function createAnimationDongle(className: string): typeof Component {
+  return class extends Component<any, any> {
+    public static displayName = 'Cycle' + className;
 
-    propTypes: {
-    },
+    constructor(props: any) {
+      super();
+      const currentValue = new Animated.Value(this.props.initialValue || 0);
+      this.state = {currentValue};
+    }
 
-    getInitialState() {
-      const currentValue = new Animated.Value(
-        this.props.initialValue || 0,
-      );
-
-      return {
-        currentValue,
-      };
-    },
-
-    componentWillReceiveProps(nextProps) {
+    public componentWillReceiveProps(nextProps: any) {
       if (nextProps.value !== this.state.currentValue._value) {
         this.runAnimation(nextProps);
       }
-    },
+    }
 
-    runAnimation({animation, options = {}, value}: any) {
-      animation(this.state.currentValue, {...options,
+    private runAnimation({animation, options = {}, value}: any) {
+      animation(this.state.currentValue, {
+        ...options,
         toValue: value,
       }).start();
-    },
+    }
 
-    render() {
+    public render() {
       const AnimatedClass = Animated[className];
       const {animate} = this.props;
 
       const animatedStyle: any = Object.keys(animate)
         .filter(key => key !== 'transform')
-        .reduce((acc, key) => (
-          {
+        .reduce(
+          (acc, key) => ({
             ...acc,
             [key]: this.state.currentValue.interpolate(animate[key]),
-          }
-        ), {});
+          }),
+          {},
+        );
 
       if (animate.transform) {
         const transforms = animate.transform.map((obj: any) => {
@@ -54,15 +49,17 @@ function createAnimationDongle(className: string) {
         animatedStyle.transform = transforms;
       }
 
-      const style = {...(this.props.style || {}), ...animatedStyle};
+      const style = {...this.props.style || {}, ...animatedStyle};
       const extraProps = {source: this.props.source};
 
-      return React.createElement(AnimatedClass, {style, ...extraProps},
+      return createElement(
+        AnimatedClass,
+        {style, ...extraProps},
         this.props.children,
       );
-    },
-  });
-};
+    }
+  };
+}
 
 export const View = createAnimationDongle('View');
 export const Text = createAnimationDongle('Text');
