@@ -3,6 +3,7 @@ import {
   ListView as RListView,
   ListViewProperties,
   ListViewDataSource,
+  createElement,
 } from 'react-native';
 
 export interface Props<T> extends Partial<ListViewProperties> {
@@ -13,35 +14,36 @@ export interface State {
   dataSource: ListViewDataSource;
 }
 
-export const ListView = React.createClass<Props<any>, State>({
-  displayName: 'CycleListView',
-  propTypes: {
-    items: React.PropTypes.array.isRequired,
-  },
+export class ListView extends React.Component<Props<any>, State> {
+  public static displayName = 'CycleListView';
+  private _listView: any;
 
-  getInitialState() {
-    const dataSource = new RListView.DataSource({rowHasChanged: (r1, r2) => r1 !== r2});
-    return {dataSource: dataSource.cloneWithRows(this.props.items)};
-  },
+  constructor(props: Props<any>) {
+    super();
+    const dataSource = new RListView.DataSource({
+      rowHasChanged: (r1, r2) => r1 !== r2,
+    });
+    this.state = {dataSource: dataSource.cloneWithRows(props.items)};
+  }
 
-  componentWillReceiveProps({items}) {
+  public componentWillReceiveProps({items}: any) {
     if (items !== this.props.items) {
       this.setState({dataSource: this.state.dataSource.cloneWithRows(items)});
     }
-  },
+  }
 
-  getScrollResponder() {
+  public getScrollResponder() {
     return this._listView.getScrollResponder();
-  },
+  }
 
-  render() {
+  public render() {
     const {items, ...listViewProps} = this.props;
-    return (
-      React.createElement(RListView, {
-        ref: listView => this._listView = listView,
-        dataSource: this.state.dataSource,
-        ...listViewProps,
-      })
-    );
-  },
-});
+    return createElement(RListView, {
+      ref: (listView: any) => {
+        this._listView = listView;
+      },
+      dataSource: this.state.dataSource,
+      ...listViewProps,
+    }) as any;
+  }
+}
